backend/requirements.txt
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
python-jose[cryptography]==3.3.0
bcrypt>=4.0
python-multipart==0.0.20
pydantic-settings>=2.0
